DMP-1168-AudioEndpointsRestructure
Merged in master
