+++ src/main/java/uk/gov/hmcts/darts/audio/component/AudioRequestResponseMapper.java
@@ -1,0 +1,8 @@
+package uk.gov.hmcts.darts.audio.component;
+
+import uk.gov.hmcts.darts.audio.entity.MediaRequestEntity;
+import uk.gov.hmcts.darts.audiorequests.model.AddAudioResponse;
+
+public interface AudioRequestResponseMapper {
+    AddAudioResponse mapToAddAudioResponse(MediaRequestEntity audioRequest);
+}
