+++ src/main/java/uk/gov/hmcts/darts/audio/component/AudioResponseMapper.java
@@ -1,7 +1,5 @@
 package uk.gov.hmcts.darts.audio.component;
 
-import uk.gov.hmcts.darts.audio.entity.MediaRequestEntity;
-import uk.gov.hmcts.darts.audio.model.AddAudioResponse;
 import uk.gov.hmcts.darts.audio.model.AudioMetadata;
 import uk.gov.hmcts.darts.common.entity.MediaEntity;
 
@@ -11,5 +9,4 @@ public interface AudioResponseMapper {
 
     List<AudioMetadata> mapToAudioMetadata(List<MediaEntity> mediaEntities);
 
-    AddAudioResponse mapToAddAudioResponse(MediaRequestEntity audioRequest);
 }
