+++ src/main/java/uk/gov/hmcts/darts/audio/controller/AudioRequestsController.java
@@ -2,15 +2,25 @@ package uk.gov.hmcts.darts.audio.controller;
 
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.core.io.InputStreamResource;
+import org.springframework.core.io.Resource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
+import uk.gov.hmcts.darts.audio.component.AudioRequestResponseMapper;
 import uk.gov.hmcts.darts.audio.component.AudioRequestSummaryMapper;
+import uk.gov.hmcts.darts.audio.entity.MediaRequestEntity;
 import uk.gov.hmcts.darts.audio.service.MediaRequestService;
 import uk.gov.hmcts.darts.audiorequests.api.AudioRequestsApi;
+import uk.gov.hmcts.darts.audiorequests.model.AddAudioResponse;
+import uk.gov.hmcts.darts.audiorequests.model.AudioRequestDetails;
 import uk.gov.hmcts.darts.audiorequests.model.AudioRequestSummary;
+import uk.gov.hmcts.darts.audit.enums.AuditActivityEnum;
+import uk.gov.hmcts.darts.audit.service.AuditService;
 import uk.gov.hmcts.darts.authorisation.annotation.Authorisation;
 
+import java.io.InputStream;
 import java.util.List;
 
 import static uk.gov.hmcts.darts.authorisation.constants.AuthorisationConstants.SECURITY_SCHEMES_BEARER_AUTH;
@@ -24,10 +34,16 @@ import static uk.gov.hmcts.darts.common.enums.SecurityRoleEnum.TRANSCRIBER;
 
 @RestController
 @RequiredArgsConstructor
+@Slf4j
 public class AudioRequestsController implements AudioRequestsApi {
 
     private final MediaRequestService mediaRequestService;
+
     private final AudioRequestSummaryMapper audioRequestSummaryMapper;
+    private final AudioRequestResponseMapper audioRequestResponseMapper;
+
+    private final AuditService auditService;
+
 
     @Override
     public ResponseEntity<List<AudioRequestSummary>> getYourAudio(Integer userId, Boolean expired) {
@@ -52,5 +68,37 @@ public class AudioRequestsController implements AudioRequestsApi {
     public ResponseEntity<Void> updateAudioRequestLastAccessedTimestamp(Integer mediaRequestId) {
         mediaRequestService.updateAudioRequestLastAccessedTimestamp(mediaRequestId);
         return new ResponseEntity<>(HttpStatus.OK);
+    }
+
+    @Override
+    @SecurityRequirement(name = SECURITY_SCHEMES_BEARER_AUTH)
+    @Authorisation(contextId = MEDIA_REQUEST_ID,
+        securityRoles = {TRANSCRIBER})
+    public ResponseEntity<Resource> download(Integer mediaRequestId) {
+        InputStream audioFileStream = mediaRequestService.download(mediaRequestId);
+
+        return new ResponseEntity<>(
+            new InputStreamResource(audioFileStream),
+            HttpStatus.OK
+        );
+    }
+
+    @Override
+    public ResponseEntity<AddAudioResponse> addAudioRequest(AudioRequestDetails audioRequestDetails) {
+        AddAudioResponse addAudioResponse;
+        MediaRequestEntity audioRequest;
+        try {
+            audioRequest = mediaRequestService.saveAudioRequest(audioRequestDetails);
+            addAudioResponse = audioRequestResponseMapper.mapToAddAudioResponse(audioRequest);
+            auditService.recordAuditRequestAudio(AuditActivityEnum.REQUEST_AUDIO,
+                                                 audioRequestDetails.getRequestor(), audioRequestDetails.getHearingId()
+            );
+        } catch (Exception e) {
+            log.error("Failed to request audio", e.getMessage());
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
+        mediaRequestService.scheduleMediaRequestPendingNotification(audioRequest);
+        return new ResponseEntity<>(addAudioResponse, HttpStatus.OK);
     }
 }
