+++ src/main/java/uk/gov/hmcts/darts/audio/service/MediaRequestService.java
@@ -3,9 +3,10 @@ package uk.gov.hmcts.darts.audio.service;
 
 import uk.gov.hmcts.darts.audio.entity.MediaRequestEntity;
 import uk.gov.hmcts.darts.audio.enums.AudioRequestStatus;
-import uk.gov.hmcts.darts.audio.model.AudioRequestDetails;
 import uk.gov.hmcts.darts.audio.service.impl.AudioRequestSummaryResult;
+import uk.gov.hmcts.darts.audiorequests.model.AudioRequestDetails;
 
+import java.io.InputStream;
 import java.util.List;
 
 public interface MediaRequestService {
@@ -25,4 +26,7 @@ public interface MediaRequestService {
     List<AudioRequestSummaryResult> viewAudioRequests(Integer userId, Boolean expired);
 
     void updateAudioRequestLastAccessedTimestamp(Integer mediaRequestId);
+
+    InputStream download(Integer mediaRequestId);
+
 }
