+++ src/main/java/uk/gov/hmcts/darts/audio/service/impl/AudioServiceImpl.java
@@ -6,14 +6,12 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.darts.audio.component.AddAudioRequestMapper;
-import uk.gov.hmcts.darts.audio.entity.MediaRequestEntity;
 import uk.gov.hmcts.darts.audio.exception.AudioApiError;
 import uk.gov.hmcts.darts.audio.model.AddAudioMetadataRequest;
 import uk.gov.hmcts.darts.audio.model.AudioFileInfo;
 import uk.gov.hmcts.darts.audio.service.AudioOperationService;
 import uk.gov.hmcts.darts.audio.service.AudioService;
 import uk.gov.hmcts.darts.audio.service.AudioTransformationService;
-import uk.gov.hmcts.darts.audit.enums.AuditActivityEnum;
 import uk.gov.hmcts.darts.audit.service.AuditService;
 import uk.gov.hmcts.darts.common.entity.HearingEntity;
 import uk.gov.hmcts.darts.common.entity.MediaEntity;
@@ -51,24 +49,6 @@ public class AudioServiceImpl implements AudioService {
             downloadPath.toFile().getAbsolutePath(),
             mediaEntity.getChannel()
         );
-    }
-
-    @Override
-    public InputStream download(Integer mediaRequestId) {
-        var transientObjectEntity = transientObjectDirectoryRepository.getTransientObjectDirectoryEntityByMediaRequest_Id(
-                mediaRequestId)
-            .orElseThrow(() -> new DartsApiException(AudioApiError.REQUESTED_DATA_CANNOT_BE_LOCATED));
-
-        UUID blobId = transientObjectEntity.getExternalLocation();
-        if (blobId == null) {
-            throw new DartsApiException(AudioApiError.REQUESTED_DATA_CANNOT_BE_LOCATED);
-        }
-        MediaRequestEntity mediaRequestEntity = transientObjectEntity.getMediaRequest();
-
-        auditService.recordAudit(AuditActivityEnum.EXPORT_AUDIO, mediaRequestEntity.getRequestor(), mediaRequestEntity.getHearing().getCourtCase());
-
-        return audioTransformationService.getOutboundAudioBlob(blobId)
-            .toStream();
     }
 
     @Override
