+++ src/main/resources/openapi/audiorequests.yaml
@@ -51,6 +51,29 @@ paths:
                 status: 400
                 detail: "Required parameter 'expired' is not present."
                 instance: "/audio-requests"
+    post:
+      tags:
+        - Audio Requests
+      summary: Users can request audio for specific cases and date/time periods
+      operationId: addAudioRequest
+      description: Adds a user audio request for processing
+      responses:
+        200:
+          description: OK
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/AddAudioResponse'
+        '400':
+          description: 'invalid input, object invalid'
+        '409':
+          description: audio request item already exists
+      requestBody:
+        content:
+          application/json:
+            schema:
+              $ref: '#/components/schemas/AudioRequestDetails'
+        description: Audio Request item to add
 
   /audio-requests/{media_request_id}:
     delete:
@@ -142,6 +165,72 @@ paths:
                 title: "The requested audio request cannot be found"
                 status: 404
 
+  /audio-requests/download:
+    get:
+      tags:
+        - Audio Requests
+      summary: Download audio
+      operationId: download
+      description: Download audio
+      parameters:
+        - in: query
+          name: media_request_id
+          required: true
+          schema:
+            type: integer
+      responses:
+        '200':
+          description: OK
+          content:
+            audio/mpeg:
+              schema:
+                type: string
+                format: binary
+            application/zip:
+              schema:
+                type: string
+                format: binary
+        '400':
+          description: Bad Request Error
+          content:
+            application/json+problem:
+              schema:
+                $ref: './problem.yaml'
+              example:
+                type: "AUTHORISATION_103"
+                title: "Failed to check authorisation for the media request"
+                status: 400
+        '401':
+          description: Unauthorised Error
+          content:
+            application/json+problem:
+              schema:
+                $ref: './problem.yaml'
+              example:
+                type: "AUTHORISATION_100"
+                title: "User is not authorised for the associated courthouse"
+                status: 401
+        '404':
+          description: Not Found Error
+          content:
+            application/json+problem:
+              schema:
+                $ref: './problem.yaml'
+              example:
+                type: "AUDIO_REQUESTS_100"
+                title: "The requested audio request cannot be found"
+                status: 404
+        '500':
+          description: The requested data cannot be located
+          content:
+            application/json+problem:
+              schema:
+                $ref: './problem.yaml'
+              example:
+                type: "AUDIO_101"
+                title: "The requested data cannot be located"
+                status: 500
+
 components:
   securitySchemes:
     bearerAuth: # arbitrary name for the security scheme
@@ -184,6 +273,82 @@ components:
           type: string
           format: date-time
           example: "2023-08-23T09:00:00Z"
+
+    defendant:
+      type: string
+      example: Joe Bloggs
+
+    AddAudioResponse:
+      type: object
+      required:
+        - request_id
+        - case_id
+        - courthouse_name
+        - defendants
+        - hearing_date
+        - start_time
+        - end_time
+      properties:
+        request_id:
+          type: integer
+          example: 1234
+        case_id:
+          type: integer
+          example: 1
+        case_number:
+          type: string
+          example: 'T4565443'
+        courthouse_name:
+          type: string
+          example: "Swansea"
+        defendants:
+          type: array
+          items:
+            $ref: '#/components/schemas/defendant'
+        hearing_date:
+          type: string
+          format: date
+          example: '10-05-2023'
+        start_time:
+          type: string
+          format: date-time
+          example: '2023-05-31T09:00:00Z'
+        end_time:
+          type: string
+          format: date-time
+          example: '2023-05-31T10:00:00Z'
+
+    AudioRequestType:
+      type: string
+      enum:
+        - DOWNLOAD
+        - PLAYBACK
+
+    AudioRequestDetails:
+      type: object
+      required:
+        - hearing_id
+        - requestor
+        - start_time
+        - end_time
+        - request_type
+      properties:
+        hearing_id:
+          type: integer
+          example: 12345
+        requestor:
+          type: integer
+          example: 4656
+        start_time:
+          type: string
+          format: date-time
+          example: '2023-05-31T09:00:00Z'
+        end_time:
+          type: string
+          format: date-time
+          example: '2023-05-31T12:00:00Z'
+        request_type:
+          $ref: '#/components/schemas/AudioRequestType'
 
     MediaRequestStatus:
       type: string
