+++ src/test/java/uk/gov/hmcts/darts/audio/component/impl/OutboundFileZipGeneratorImplTest.java
@@ -40,7 +40,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.darts.audio.enums.AudioRequestStatus.OPEN;
-import static uk.gov.hmcts.darts.audio.model.AudioRequestType.DOWNLOAD;
+import static uk.gov.hmcts.darts.audiorequests.model.AudioRequestType.DOWNLOAD;
 
 @SuppressWarnings({"PMD.AvoidThrowingRawExceptionTypes", "PMD.ExcessiveImports"})
 @ExtendWith(MockitoExtension.class)
