+++ src/test/java/uk/gov/hmcts/darts/audio/service/impl/MediaRequestServiceImplTest.java
@@ -1,5 +1,6 @@
 package uk.gov.hmcts.darts.audio.service.impl;
 
+import com.azure.core.util.BinaryData;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -9,12 +10,16 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.darts.audio.entity.MediaRequestEntity;
-import uk.gov.hmcts.darts.audio.model.AudioRequestDetails;
+import uk.gov.hmcts.darts.audio.exception.AudioApiError;
 import uk.gov.hmcts.darts.audio.repository.MediaRequestRepository;
+import uk.gov.hmcts.darts.audiorequests.model.AudioRequestDetails;
+import uk.gov.hmcts.darts.audit.service.AuditService;
 import uk.gov.hmcts.darts.common.entity.CourtCaseEntity;
 import uk.gov.hmcts.darts.common.entity.HearingEntity;
+import uk.gov.hmcts.darts.common.entity.MediaEntity;
 import uk.gov.hmcts.darts.common.entity.TransientObjectDirectoryEntity;
 import uk.gov.hmcts.darts.common.entity.UserAccountEntity;
+import uk.gov.hmcts.darts.common.exception.DartsApiException;
 import uk.gov.hmcts.darts.common.repository.HearingRepository;
 import uk.gov.hmcts.darts.common.repository.TransientObjectDirectoryRepository;
 import uk.gov.hmcts.darts.common.repository.UserAccountRepository;
@@ -22,24 +27,34 @@ import uk.gov.hmcts.darts.datamanagement.api.impl.DataManagementApiImpl;
 import uk.gov.hmcts.darts.notification.api.NotificationApi;
 import uk.gov.hmcts.darts.notification.dto.SaveNotificationToDbRequest;
 
+import java.io.IOException;
+import java.io.InputStream;
 import java.time.OffsetDateTime;
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.any;
+import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.darts.audio.enums.AudioRequestStatus.OPEN;
-import static uk.gov.hmcts.darts.audio.model.AudioRequestType.DOWNLOAD;
+import static uk.gov.hmcts.darts.audiorequests.model.AudioRequestType.DOWNLOAD;
 
 @ExtendWith(MockitoExtension.class)
+@SuppressWarnings({"PMD.ExcessiveImports"})
 class MediaRequestServiceImplTest {
 
     private static final Integer TEST_REQUESTER = 1234;
     private static final String OFFSET_T_09_00_00_Z = "2023-05-31T09:00:00Z";
     private static final String OFFSET_T_12_00_00_Z = "2023-05-31T12:00:00Z";
+    private static final String DUMMY_FILE_CONTENT = "DUMMY FILE CONTENT";
+    private static final OffsetDateTime START_TIME = OffsetDateTime.parse("2023-01-01T12:00:00Z");
+    private static final OffsetDateTime END_TIME = OffsetDateTime.parse("2023-01-01T13:00:00Z");
+
 
     @InjectMocks
     private MediaRequestServiceImpl mediaRequestService;
@@ -56,6 +71,8 @@ class MediaRequestServiceImplTest {
     private DataManagementApiImpl dataManagementApi;
     @Mock
     private NotificationApi notificationApi;
+    @Mock
+    private AuditService auditService;
 
     private HearingEntity mockHearingEntity;
     private MediaRequestEntity mockMediaRequestEntity;
@@ -174,4 +191,73 @@ class MediaRequestServiceImplTest {
         verify(transientObjectDirectoryRepository, Mockito.times(0)).deleteById(any());
     }
 
+    @Test
+    void downloadShouldReturnExpectedData() throws IOException {
+        MediaEntity mediaEntity = new MediaEntity();
+        mediaEntity.setId(1);
+        mediaEntity.setStart(START_TIME);
+        mediaEntity.setEnd(END_TIME);
+        mediaEntity.setChannel(1);
+
+        UserAccountEntity userAccountEntity = new UserAccountEntity();
+        userAccountEntity.setId(1);
+        CourtCaseEntity courtCaseEntity = new CourtCaseEntity();
+        courtCaseEntity.setId(1);
+        HearingEntity hearingEntity = new HearingEntity();
+        hearingEntity.setId(1);
+        courtCaseEntity.setHearings(List.of(hearingEntity));
+        MediaRequestEntity mediaRequestEntity = new MediaRequestEntity();
+        mediaRequestEntity.setId(1);
+        mediaRequestEntity.setRequestor(userAccountEntity);
+        mediaRequestEntity.setHearing(hearingEntity);
+        var blobUuid = UUID.randomUUID();
+        var transientObjectDirectoryEntity = new TransientObjectDirectoryEntity();
+        transientObjectDirectoryEntity.setExternalLocation(blobUuid);
+        transientObjectDirectoryEntity.setMediaRequest(mediaRequestEntity);
+
+        var mediaRequestId = 1;
+        when(transientObjectDirectoryRepository.getTransientObjectDirectoryEntityByMediaRequest_Id(mediaRequestId))
+            .thenReturn(Optional.of(transientObjectDirectoryEntity));
+        //when(auditService.recordAudit(any(), any(), any()))
+        doNothing().when(auditService).recordAudit(any(), any(), any());
+
+        when(dataManagementApi.getBlobDataFromOutboundContainer(blobUuid))
+            .thenReturn(BinaryData.fromBytes(DUMMY_FILE_CONTENT.getBytes()));
+
+        try (InputStream inputStream = mediaRequestService.download(mediaRequestId)) {
+            byte[] bytes = inputStream.readAllBytes();
+            assertEquals(DUMMY_FILE_CONTENT, new String(bytes));
+        }
+    }
+
+    @Test
+    void downloadShouldThrowExceptionWhenRelatedTransientObjectCannotBeFound() {
+        var mediaRequestId = 1;
+        when(transientObjectDirectoryRepository.getTransientObjectDirectoryEntityByMediaRequest_Id(mediaRequestId))
+            .thenReturn(Optional.empty());
+
+        var exception = assertThrows(
+            DartsApiException.class,
+            () -> mediaRequestService.download(mediaRequestId)
+        );
+
+        assertEquals(AudioApiError.REQUESTED_DATA_CANNOT_BE_LOCATED, exception.getError());
+    }
+
+    @Test
+    void downloadShouldThrowExceptionWhenTransientObjectHasNoExternalLocationValue() {
+        var transientObjectDirectoryEntity = new TransientObjectDirectoryEntity();
+        transientObjectDirectoryEntity.setExternalLocation(null);
+
+        var mediaRequestId = 1;
+        when(transientObjectDirectoryRepository.getTransientObjectDirectoryEntityByMediaRequest_Id(mediaRequestId))
+            .thenReturn(Optional.of(transientObjectDirectoryEntity));
+
+        var exception = assertThrows(
+            DartsApiException.class,
+            () -> mediaRequestService.download(mediaRequestId)
+        );
+
+        assertEquals(AudioApiError.REQUESTED_DATA_CANNOT_BE_LOCATED, exception.getError());
+    }
 }
